+++ src/elements/flip-card/flip-card-details/flip-card-details.stories.ts
@@ -1,29 +1,24 @@
-import type { Decorator, Meta, StoryObj } from '@storybook/web-components-vite';
+import type { Meta, StoryObj } from '@storybook/web-components-vite';
 import type { TemplateResult } from 'lit';
 import { html } from 'lit';
-import { withActions } from 'storybook/actions/decorator';
 
 import readme from './readme.md?raw';
 
-import '../../link.js';
-import './flip-card-details.component.js';
+import '../../card.js';
 
-const Template = (): TemplateResult =>
-  html`<sbb-flip-card-details style="--sbb-flip-card-details-opacity: 1">
-    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam luctus ornare condimentum. Vivamus
-    turpis elit, dapibus eget fringilla pellentesque, lobortis in nibh. Duis dapibus vitae tortor
-    ullamcorper maximus. In convallis consectetur felis.
-    <sbb-link href="https://www.sbb.ch" negative>Link</sbb-link>
-  </sbb-flip-card-details>`;
+const Template = (): TemplateResult => html`
+  <sbb-card color="milk">
+    'sbb-flip-card-details' is an element to be only used together with 'sbb-flip-card'. See
+    'sbb-flip-card' examples to see it in action.
+  </sbb-card>
+`;
 
 export const Default: StoryObj = {
   render: Template,
 };
 
 const meta: Meta = {
-  decorators: [withActions as Decorator],
   parameters: {
-    backgroundColor: () => 'var(--sbb-color-midnight)',
     docs: {
       extractComponentDescription: () => readme,
     },