+++ src/elements/form-field/form-field-clear/form-field-clear.visual.spec.ts
@@ -39,7 +39,7 @@ describe(`sbb-form-field-clear`, () => {
               <sbb-form-field-clear></sbb-form-field-clear>
             </sbb-form-field>
           `,
-          { backgroundColor: negative ? 'var(--sbb-color-charcoal)' : undefined },
+          { backgroundColor: negative ? 'var(--sbb-background-color-1-negative)' : undefined },
         );
       });