+++ src/elements/menu/menu/__snapshots__/menu.snapshot.spec.snap.js
@@ -15,6 +15,7 @@ snapshots["sbb-menu renders DOM"] =
     data-slot-names="unnamed"
     href="https://www.sbb.ch/en"
     icon-placement="start"
+    negative=""
     size="s"
   >
     Profile
@@ -69,6 +70,18 @@ snapshots["sbb-menu renders Shadow DOM"] =
     <div class="sbb-menu__content">
       <slot>
       </slot>
+      <sbb-divider orientation="horizontal">
+      </sbb-divider>
+      <sbb-menu-button
+        data-action=""
+        data-button=""
+        data-slot-names="unnamed"
+        icon-name="chevron-small-left-small"
+        id="sbb-menu__back-button"
+        tabindex="0"
+      >
+        Go back
+      </sbb-menu-button>
     </div>
   </div>
 </div>
@@ -154,6 +167,18 @@ snapshots["sbb-menu renders with list Shadow DOM"] =
         <slot>
         </slot>
       </span>
+      <sbb-divider orientation="horizontal">
+      </sbb-divider>
+      <sbb-menu-button
+        data-action=""
+        data-button=""
+        data-slot-names="unnamed"
+        icon-name="chevron-small-left-small"
+        id="sbb-menu__back-button"
+        tabindex="0"
+      >
+        Go back
+      </sbb-menu-button>
     </div>
   </div>
 </div>
@@ -217,6 +242,7 @@ snapshots["sbb-menu renders open DOM"] =
     data-slot-names="unnamed"
     href="https://www.sbb.ch/en"
     icon-placement="start"
+    negative=""
     size="s"
   >
     Profile
@@ -271,6 +297,18 @@ snapshots["sbb-menu renders open Shadow DOM"] =
     <div class="sbb-menu__content">
       <slot>
       </slot>
+      <sbb-divider orientation="horizontal">
+      </sbb-divider>
+      <sbb-menu-button
+        data-action=""
+        data-button=""
+        data-slot-names="unnamed"
+        icon-name="chevron-small-left-small"
+        id="sbb-menu__back-button"
+        tabindex="0"
+      >
+        Go back
+      </sbb-menu-button>
     </div>
   </div>
 </div>