+++ src/elements/menu/menu/menu.visual.spec.ts
@@ -1,9 +1,10 @@
+import { sendKeys } from '@web/test-runner-commands';
 import { html } from 'lit';
 import { repeat } from 'lit/directives/repeat.js';
 import { styleMap } from 'lit/directives/style-map.js';
 
 import type { SbbButtonElement } from '../../button/button.js';
-import { describeViewports, visualDiffDefault } from '../../core/testing/private.js';
+import { describeViewports, tabKey, visualDiffDefault } from '../../core/testing/private.js';
 
 import '../../button/button.js';
 import '../../divider.js';
@@ -11,6 +12,7 @@ import '../../link.js';
 import '../menu-link.js';
 import '../menu-button.js';
 import './menu.component.js';
+import type { SbbMenuElement } from './menu.component.js';
 
 describe(`sbb-menu`, () => {
   const userNameStyle = {
@@ -20,7 +22,7 @@ describe(`sbb-menu`, () => {
   };
 
   const userInfoStyle = {
-    color: 'var(--sbb-color-graphite)',
+    color: 'light-dark(var(--sbb-color-graphite), var(--sbb-color-granite))',
     fontSize: 'var(--sbb-font-size-text-xxs)',
   };
 
@@ -45,9 +47,12 @@ describe(`sbb-menu`, () => {
           `,
           { minHeight: '400px' },
         );
-        setup.withPostSetupAction(() => {
+        setup.withPostSetupAction(async () => {
           const button = setup.snapshotElement.querySelector<SbbButtonElement>('#menu-trigger-1')!;
           button.click();
+
+          // Test focus outline
+          await sendKeys({ press: tabKey });
         });
       }),
     );
@@ -83,9 +88,7 @@ describe(`sbb-menu`, () => {
             <sbb-menu trigger="menu-trigger-3">
               <div style=${styleMap(userNameStyle)}>Christina Müller</div>
               <span style=${styleMap(userInfoStyle)}>UIS9057</span>
-              <sbb-block-link href="https://www.sbb.ch/en" negative size="xs">
-                Profile
-              </sbb-block-link>
+              <sbb-block-link href="https://www.sbb.ch/en" size="xs"> Profile </sbb-block-link>
               <sbb-divider></sbb-divider>
               <sbb-menu-button icon-name="swisspass-small">
                 Very long label that exceeds the maximum width of the menu
@@ -103,6 +106,121 @@ describe(`sbb-menu`, () => {
 
         setup.withPostSetupAction(() => {
           const button = setup.snapshotElement.querySelector<SbbButtonElement>('#menu-trigger-3')!;
+          button.click();
+        });
+      }),
+    );
+
+    it(
+      'nested default',
+      visualDiffDefault.with(async (setup) => {
+        await setup.withFixture(
+          html`
+            <sbb-button id="menu-trigger-1" size="m">Menu trigger</sbb-button>
+            <sbb-menu trigger="menu-trigger-1">
+              <sbb-menu-button icon-name="link-small">First level menu</sbb-menu-button>
+              <sbb-menu-button icon-name="pen-small">First level menu</sbb-menu-button>
+              <sbb-menu-button icon-name="swisspass-small" id="menu-trigger-2">
+                First level menu
+              </sbb-menu-button>
+              <sbb-menu-button icon-name="cross-small">First level menu</sbb-menu-button>
+            </sbb-menu>
+            <sbb-menu trigger="menu-trigger-2">
+              <sbb-menu-button icon-name="link-small">Second level menu</sbb-menu-button>
+              <sbb-menu-button icon-name="pen-small">Second level menu</sbb-menu-button>
+              <sbb-menu-button icon-name="swisspass-small">Second level menu</sbb-menu-button>
+              <sbb-menu-button icon-name="cross-small">Second level menu</sbb-menu-button>
+            </sbb-menu>
+          `,
+          { minHeight: '400px' },
+        );
+
+        setup.withPostSetupAction(() => {
+          let button = setup.snapshotElement.querySelector<SbbButtonElement>('#menu-trigger-1')!;
+          const menu = setup.snapshotElement.querySelector<SbbMenuElement>('sbb-menu')!;
+
+          button.click();
+
+          button = menu.querySelector('#menu-trigger-2')!;
+          button.click();
+        });
+      }),
+    );
+  });
+
+  describeViewports({ viewports: ['medium'], viewportHeight: 500 }, () => {
+    it(
+      'nested wrap left',
+      visualDiffDefault.with(async (setup) => {
+        await setup.withFixture(
+          html`
+            <div style="display: flex;">
+              <sbb-button id="menu-trigger-1" size="m" style="margin-inline-start: auto;"
+                >Menu trigger</sbb-button
+              >
+              <sbb-menu trigger="menu-trigger-1">
+                <sbb-menu-button icon-name="link-small">First level menu</sbb-menu-button>
+                <sbb-menu-button icon-name="pen-small">First level menu</sbb-menu-button>
+                <sbb-menu-button icon-name="swisspass-small" id="menu-trigger-2">
+                  First level menu
+                </sbb-menu-button>
+                <sbb-menu-button icon-name="cross-small">First level menu</sbb-menu-button>
+              </sbb-menu>
+              <sbb-menu trigger="menu-trigger-2">
+                <sbb-menu-button icon-name="link-small">Second level menu</sbb-menu-button>
+                <sbb-menu-button icon-name="pen-small">Second level menu</sbb-menu-button>
+                <sbb-menu-button icon-name="swisspass-small">Second level menu</sbb-menu-button>
+                <sbb-menu-button icon-name="cross-small">Second level menu</sbb-menu-button>
+              </sbb-menu>
+            </div>
+          `,
+          { minHeight: '500px' },
+        );
+
+        setup.withPostSetupAction(() => {
+          let button = setup.snapshotElement.querySelector<SbbButtonElement>('#menu-trigger-1')!;
+          const menu = setup.snapshotElement.querySelector<SbbMenuElement>('sbb-menu')!;
+
+          button.click();
+
+          button = menu.querySelector('#menu-trigger-2')!;
+          button.click();
+        });
+      }),
+    );
+  });
+  describeViewports({ viewports: ['medium'], viewportHeight: 300 }, () => {
+    it(
+      'nested short page',
+      visualDiffDefault.with(async (setup) => {
+        await setup.withFixture(
+          html`
+            <sbb-button id="menu-trigger-1" size="m">Menu trigger</sbb-button>
+            <sbb-menu trigger="menu-trigger-1">
+              <sbb-menu-button icon-name="link-small">First level menu</sbb-menu-button>
+              <sbb-menu-button icon-name="pen-small">First level menu</sbb-menu-button>
+              <sbb-menu-button icon-name="swisspass-small" id="menu-trigger-2">
+                First level menu
+              </sbb-menu-button>
+              <sbb-menu-button icon-name="cross-small">First level menu</sbb-menu-button>
+            </sbb-menu>
+            <sbb-menu trigger="menu-trigger-2">
+              <sbb-menu-button icon-name="link-small">Second level menu</sbb-menu-button>
+              <sbb-menu-button icon-name="pen-small">Second level menu</sbb-menu-button>
+              <sbb-menu-button icon-name="swisspass-small">Second level menu</sbb-menu-button>
+              <sbb-menu-button icon-name="cross-small">Second level menu</sbb-menu-button>
+            </sbb-menu>
+          `,
+          { minHeight: '300px' },
+        );
+
+        setup.withPostSetupAction(() => {
+          let button = setup.snapshotElement.querySelector<SbbButtonElement>('#menu-trigger-1')!;
+          const menu = setup.snapshotElement.querySelector<SbbMenuElement>('sbb-menu')!;
+
+          button.click();
+
+          button = menu.querySelector('#menu-trigger-2')!;
           button.click();
         });
       }),