+++ src/elements/paginator/compact-paginator/compact-paginator.stories.ts
@@ -157,7 +157,9 @@ const meta: Meta = {
       handles: [SbbCompactPaginatorElement.events.page],
     },
     backgroundColor: (context: StoryContext) =>
-      context.args.negative ? 'var(--sbb-color-black)' : 'var(--sbb-color-white)',
+      context.args.negative
+        ? 'var(--sbb-background-color-1-negative)'
+        : 'var(--sbb-background-color-1)',
     docs: {
       extractComponentDescription: () => readme,
     },